추시 d-day 정상 출력 및 정렬

+++ app/src/main/java/com/example/younghyeon/test0901/FoodInputActivity.java
@@ -236,7 +236,7 @@ public class FoodInputActivity extends Activity {
                         intent.putExtra("num", num);
                         intent.putExtra("image_num", image_num);
                         intent.putExtra("position", position);
-                        intent.putExtra("d_day", diff);
+                        intent.putExtra("d_day", (int)diff);
